% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_left.R
\name{align_left}
\alias{align_left}
\title{Left-align text and pad to same length.}
\usage{
align_left(
  text,
  width = max(nchar(text)),
  type = c("non-trimmed", "trimmed"),
  sep = " ",
  keep.empty = TRUE
)
}
\arguments{
\item{text}{\code{character} with strings to be left-aligned.}

\item{width}{\code{numeric} indicating the desired length of strings.}

\item{type}{\code{character} specifying pre-processing. Trims whitespace from \code{text} if set to "trimmed".}

\item{sep}{\code{character} specifying the padding. May give unexpected results if not a single character.}

\item{keep.empty}{\code{logical} Indicate if empty strings should be kept empty.}
}
\value{
\code{character} with left-aligned text.
}
\description{
Left-align text and pad to same length.
}
