% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approvedpkgs.R
\name{check_approved}
\alias{check_approved}
\title{Add Approved column on session packages tibble}
\usage{
check_approved(approved_pkg_loc, session_pkgs, output_file = NULL)
}
\arguments{
\item{approved_pkg_loc}{The library location where the approved packages are}

\item{session_pkgs}{The loaded packages in the current session}

\item{output_file}{String. Name of file where the modified tibble will be saved.}
}
\value{
session_pkgs with additional column of Approved indicating if the package is approved or not
}
\description{
A utility function to help you build your approved packages .
}
\examples{
approved_pkgs <- c( "/opt/R/4.2.0/lib/R/library", "/opt/R/4.2.0/lib/R/site-library" )
# build and return
check_approved(approved_pkgs, sessioninfo::package_info())

# build and save
dir <- tempdir()
check_approved(approved_pkgs, sessioninfo::package_info(), file.path(dir, "approved.rds"))

}
