% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumE.R
\name{cumE}
\alias{cumE}
\title{Compute the accumulated cyclone energy of a given storm}
\usage{
cumE(storm_data, storm_id)
}
\arguments{
\item{storm_data}{| the dataframe for the hurricane information}

\item{storm_id}{| the storm IDs to be plot for its tracks}
}
\value{
energy | accumulated cyclone energy of the storm
}
\description{
Compute the accumulated cyclone energy of a given storm
}
\examples{
data(processed_hurdat)
new = cumE(processed_hurdat,"AL182019")
}
