% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_logging.R
\name{custom_logging}
\alias{custom_logging}
\alias{log_read}
\alias{log_write}
\alias{log_delete}
\title{Helper function to log custom messages}
\usage{
log_read(file, log = Sys.getenv("WHIRL_LOG_MSG"))

log_write(file, log = Sys.getenv("WHIRL_LOG_MSG"))

log_delete(file, log = Sys.getenv("WHIRL_LOG_MSG"))
}
\description{
Useful for e.g. read and write operations on databases etc.
that are not automatically captured.
}
