% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_by_config.R
\name{define_paths}
\alias{define_paths}
\title{Find path for files to log}
\usage{
define_paths(step, root_dir, cli_level = cli::cli_h1)
}
\arguments{
\item{step}{a step from yaml (as a list in R)}

\item{root_dir}{the root dir of the yaml file}

\item{cli_level}{Have to be a function from cli}
}
\description{
Find path for files to log
}
