% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_input.R
\name{enrich_input}
\alias{enrich_input}
\title{Converting input to a standardized list with that contain step names and
files paths}
\usage{
enrich_input(input, steps = NULL)
}
\arguments{
\item{input}{Can be a vector, a list or a whirl config file.}

\item{steps}{A filter argument for selecting specific steps that should be
executed}
}
\value{
A list
}
\description{
Converting input to a standardized list with that contain step names and
files paths
}
