% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.R
\name{interpolate}
\alias{interpolate}
\title{A function for interpolating a storm track to 30 min increments.}
\usage{
interpolate(storm_data, storm_id)
}
\arguments{
\item{storm_data}{| the dataframe for the hurricane information}

\item{storm_id}{| the storm IDs to be plot for its tracks}
}
\value{
new | interpolated data.frame
}
\description{
A function for interpolating a storm track to 30 min increments.
}
\examples{
data(processed_hurdat)
new = interpolate(processed_hurdat,"AL021857")
}
