% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landfall.R
\name{landfall}
\alias{landfall}
\title{Determine whether storm made landfall within US continent.}
\usage{
landfall(storm_data, storm_id)
}
\arguments{
\item{storm_data}{| the dataframe for the hurricane information}

\item{storm_id}{| the storm IDs to be plot for its tracks}
}
\value{
logic value | whether it landfall or not
}
\description{
Determine whether storm made landfall within US continent.
}
\examples{
data(processed_hurdat)
logic.landfall = landfall(processed_hurdat,"AL061851")

}
