% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exuction_sequence.R
\name{logs_from_whirl_config}
\alias{logs_from_whirl_config}
\title{Create logs for files from a yaml configuration}
\usage{
logs_from_whirl_config(file, summary_dir = ".")
}
\arguments{
\item{file}{yaml configuration file}

\item{summary_dir}{A character string of file path specifying the directory where the summary log will be stored.}
}
\value{
A list containing the execution results for each script. Each element of the
list is a character string indicating the success or failure of the script execution.
}
\description{
Create logs for files from a yaml configuration
}
\examples{
file_ <- system.file(
            "examples",
            "sequence_exuction",
             "_whirl.yaml",
             package = "whirl"
             )
logs_from_whirl_config(file_, tempdir())
}
