% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_with_base.R
\name{normalize_with_base}
\alias{normalize_with_base}
\title{Normalize a Path with Respect to a Base Directory}
\usage{
normalize_with_base(path, base = ".")
}
\arguments{
\item{path}{A character string representing the path to be normalized. Can be relative, absolute, or start with \code{~}.}

\item{base}{A character string representing the base directory with respect to which the path should be normalized.
The default is the current working directory (".").}
}
\value{
A character string representing the normalized path.
}
\description{
This function normalizes a given path with respect to a specified base directory.
If the path is relative, it combines the base directory and the path, then normalizes the resulting path.
If the path is absolute or starts with \code{~}, it normalizes the path directly.
If no base directory is specified, the current working directory is used as the base.
}
\examples{
\dontrun{
base <- "/my/base/directory"
relative_path <- "subdir/file.txt"
normalized_path <- normalize_with_base(relative_path, base)
print(normalized_path)

# Using the current working directory as the base
relative_path <- "subdir/file.txt"
normalized_path <- normalize_with_base(relative_path)
print(normalized_path)

# Using an absolute path
absolute_path <- "/another/directory/subdir/file.txt"
normalized_path <- normalize_with_base(absolute_path)
print(normalized_path)

# Using a path starting with ~
home_path <- "~/subdir/file.txt"
normalized_path <- normalize_with_base(home_path)
print(normalized_path)
}
}
\keyword{internal}
