% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whirl-package.R
\name{options}
\alias{options}
\title{whirl Options}
\description{
Internally used, package-specific options. All options will prioritize R options() values, and fall back to environment variables if undefined. If neither the option nor the environment variable is set, a default value is used.
}
\section{Checking Option Values}{

Option values specific to \code{whirl} can be
accessed by passing the package name to \code{env}.

\if{html}{\out{<div class="sourceCode">}}\preformatted{options::opts(env = "whirl")

options::opt(x, default, env = "whirl")
}\if{html}{\out{</div>}}
}

\section{Options}{

\describe{
\item{track_files}{\describe{
track files\item{default: }{\preformatted{FALSE}}
\item{option: }{whirl.track_files}
\item{envvar: }{R_WHIRL_TRACK_FILES (evaluated if possible, raw string otherwise)}
}}

\item{track_files_discards}{\describe{
strace discards keywords to use to discard not required lines\item{default: }{\preformatted{""}}
\item{option: }{whirl.track_files_discards}
\item{envvar: }{R_WHIRL_TRACK_FILES_DISCARDS (evaluated if possible, raw string otherwise)}
}}

\item{check_renv}{\describe{
renv\item{default: }{\preformatted{FALSE}}
\item{option: }{whirl.check_renv}
\item{envvar: }{R_WHIRL_CHECK_RENV (evaluated if possible, raw string otherwise)}
}}

}
}

\seealso{
options getOption Sys.setenv Sys.getenv
}
