% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{options_params}
\alias{options_params}
\title{Internal reuse of options description}
\arguments{
\item{approved_pkgs_folder}{Approved folder library packages
Default: NULL
Option: approved_pkgs_folder
Environment variable: R_APPROVED_PKGS_FOLDER
@param approved_pkgs_url Approved URL library packages
Default: NULL
Option: approved_pkgs_url
Environment variable: R_APPROVED_PKGS_URL
@param check_renv Should the projects renv status be checked?
Default: FALSE
Option: check_renv
Environment variable: R_CHECK_RENV
@param n_workers Number of simultanous workers used in the run function. A maximum of 128 workers is allowed.
Default: 1
Option: n_workers
Environment variable: R_N_WORKERS
@param out_formats Which log format(s) to produce. Possiblities are \code{html}, \code{json}, and markdown formats:\code{gfm}, \code{commonmark}, and \code{markua}.
Default: "html"
Option: out_formats
Environment variable: R_OUT_FORMATS
@param track_files Should files read and written be tracked? Currently only supported on Linux.
Default: FALSE
Option: track_files
Environment variable: R_TRACK_FILES
@param track_files_discards List of file naming patterns not be tracked when track_files = TRUE
Default: c("^/lib", "^/etc", "^/lib64", "^/usr", "^/var", "^/opt", "^/sys",
Default: "^/proc", "^/tmp", "^/null", "^/urandom", "^/.cache", "C:/Users/lgaka/AppData/Local/R/win-library/4.4",
Default: "C:/Program Files/R/R-4.4.2/library")
Option: track_files_discards
Environment variable: R_TRACK_FILES_DISCARDS
@param track_files_keep List of file naming patterns alway to be tracked when track_files = TRUE
Default: "^C:/Users/lgaka/Downloads/whirl"
Option: track_files_keep
Environment variable: R_TRACK_FILES_KEEP
@param verbosity_level How chatty should the log be? Possibilities are \code{quiet}, \code{minimal} and \code{verbose}.
Default: "verbose"
Option: verbosity_level
Environment variable: R_VERBOSITY_LEVEL}
}
\description{
Internal reuse of options description
}
\keyword{internal}
