% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_storm_size_helper.R
\name{plot_storm_size_helper}
\alias{plot_storm_size_helper}
\title{Helper function to help with plot_storm_size function to get the
row and track of a specified storm year and name}
\usage{
plot_storm_size_helper(storm_data, storm_name, storm_year)
}
\arguments{
\item{storm_data}{| processed hur data, dataframe}

\item{storm_name}{| a stirng with specified storm name}

\item{storm_year}{| a stirng with specified storm year}
}
\value{
df | dataframe that contains the strongest landfall row
}
\description{
Helper function to help with plot_storm_size function to get the
row and track of a specified storm year and name
}
\examples{
data(processed_hurdat)
katrina = plot_storm_size_helper(processed_hurdat, "katrina", "2005")
}
