% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto.R
\name{quarto_callout}
\alias{quarto_callout}
\title{Create a Quarto callout block with code}
\usage{
quarto_callout(
  text = NULL,
  title = NULL,
  type = c("note", "warning", "important", "tip", "caution"),
  collapse = NULL
)
}
\arguments{
\item{text}{description}

\item{title}{description}

\item{type}{description}

\item{collapse}{description}
}
\description{
See https://quarto.org/docs/authoring/callouts.html
}
