% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_regexp.R
\name{read_regexp}
\alias{read_regexp}
\title{Fetch files if path uses regular expression}
\usage{
read_regexp(input)
}
\arguments{
\item{input}{A character vector with paths to files that should be executed.}
}
\value{
A character vector with the paths to the files. If regexp have been
used as input these will be solved to the actual files matching the
criteria.
}
\description{
Fetch files if path uses regular expression
}
