% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strace.R
\name{read_strace}
\alias{read_strace}
\title{Read STRACE file}
\usage{
read_strace(
  path,
  strace_discards = c("^/lib", "^/etc", "^/lib64", "^/usr", "^/var", "^/opt", "^/sys",
    "^/proc", "^/tmp", "^.$", paste0("^", .libPaths()))
)
}
\arguments{
\item{path}{Path to the .strace file}

\item{strace_discards}{characters to identify records to discard}
}
\value{
A \code{tibble} with strace information.
}
\description{
Read STRACE file
}
