% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strace.R
\name{readstrace_info}
\alias{readstrace_info}
\title{Get session info}
\usage{
readstrace_info(path, strace_discards = NULL)
}
\arguments{
\item{path}{a character vector with path name}

\item{strace_discards}{keywords to use to discard not required lines}
}
\description{
Retrieve session info and add quarto info if not already there
}
