% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strace.R
\name{refine_strace}
\alias{refine_strace}
\title{refine strace output}
\usage{
refine_strace(data_strace)
}
\arguments{
\item{data_strace}{\itemize{
\item file lines
}}
}
\value{
tibble
}
\description{
refine strace output
}
