% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_summary.R
\name{render_summary}
\alias{render_summary}
\title{Render dataframe into a summary.html file}
\usage{
render_summary(input, summary_file = "summary.html")
}
\arguments{
\item{input}{The input data.frame that should be rendered into a summary.html file}

\item{summary_file}{A character string specifying the path where the summary HTML file should be saved. Defaults to \code{"summary.html"}.}
}
\value{
Takes a dataframe as input and returns a log in html format
}
\description{
Render dataframe into a summary.html file
}
