% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{retrieve_fpath}
\alias{retrieve_fpath}
\title{Retrieve path of file from folder/subfolders}
\usage{
retrieve_fpath(in_file)
}
\arguments{
\item{in_file}{\itemize{
\item file to search path for
}}
}
\value{
A path to a file
}
\description{
Retrieve path of file from folder/subfolders
}
\examples{
retrieve_fpath("prg1.R")
}
