% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_by_config.R
\name{run_by_config}
\alias{run_by_config}
\title{Create logs for files from a yaml configuration}
\usage{
run_by_config(file, steps, summary_dir = ".", root_dir = dirname(file))
}
\arguments{
\item{file}{yaml configuration file}

\item{steps}{An optional argument that can be used if only certain steps
within a config files is to be executed.}

\item{summary_dir}{A character string of file path specifying the directory
where the summary log will be stored.}

\item{root_dir}{By default, the root dir of the yaml file.}
}
\value{
A list containing the execution results for each script. Each element
of the list is a character string indicating the success or failure of the
script execution.
}
\description{
Create logs for files from a yaml configuration
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

file_ <- system.file(
  "examples",
  "sequence_exuction",
  "_whirl.yaml",
  package = "whirl"
)
run_by_config(file_, tempdir())
\dontshow{\}) # examplesIf}
}
