% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{run_script}
\alias{run_script}
\title{Run script
test with a single script}
\usage{
run_script(
  script,
  track_files = FALSE,
  strace_discards = NULL,
  renv = FALSE,
  output_format = c("html"),
  out_dir = dirname(script)
)
}
\arguments{
\item{script}{path}

\item{track_files}{\code{logical}}

\item{strace_discards}{keywords to use to discard not required lines}

\item{renv}{\code{logical}}

\item{output_format}{\code{vector} either html, gfm, commonmark or markua format(s) of the log output(s)}

\item{out_dir}{description}
}
\description{
Run script
test with a single script
}
