% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_whirl.R
\name{use_biocompute}
\alias{use_biocompute}
\title{Use whirl to create biocompute logs}
\usage{
use_biocompute(
  config_file = "_whirl.yml",
  parametrics_file = "_parametrics.yml"
)
}
\arguments{
\item{config_file}{Path to the whirl config file, relative to the project}

\item{parametrics_file}{Path to the biocompute parametrics file, relative to the project}
}
\description{
Utility function to setup execution with whirl in your project suitable for
creating biocompute logs with \code{write_biocompute()}:
\enumerate{
\item Creates configuration file (default \verb{_whirl.yml}) with default values for the \code{biocompute} metadata.
\item Updates \code{.gitignore} to not include log files
}

See \code{vignette("whirl")} for how to specify paths inside the
configuration file.
}
