% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_whirl.R
\name{use_whirl}
\alias{use_whirl}
\title{Use whirl}
\usage{
use_whirl(config_file = "_whirl.yml")
}
\arguments{
\item{config_file}{Path to the whirl config file, relative to the project}
}
\description{
Utility function to setup execution with whirl in your project:
\enumerate{
\item Creates configuration file (default \verb{_whirl.yml})
\item Updates \code{.gitignore} to not include log files
}

See \code{vignette("whirl")} for how to specify paths inside the
configuration file.
}
