% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whirl_r_session.R
\name{whirl_r_session}
\alias{whirl_r_session}
\title{Internal whirl extention of the callr::r_session class to be used in \code{run_script()}}
\description{
Internal whirl extention of the callr::r_session class to be used in \code{run_script()}

Internal whirl extention of the callr::r_session class to be used in \code{run_script()}
}
\section{Super classes}{
\code{\link[processx:process]{processx::process}} -> \code{\link[callr:r_session]{callr::r_session}} -> \code{whirl_r_session}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-whirl_r_session-new}{\code{whirl_r_session$new()}}
\item \href{#method-whirl_r_session-finalize}{\code{whirl_r_session$finalize()}}
\item \href{#method-whirl_r_session-print}{\code{whirl_r_session$print()}}
\item \href{#method-whirl_r_session-get_dir}{\code{whirl_r_session$get_dir()}}
\item \href{#method-whirl_r_session-start_strace}{\code{whirl_r_session$start_strace()}}
\item \href{#method-whirl_r_session-run_something}{\code{whirl_r_session$run_something()}}
\item \href{#method-whirl_r_session-clone}{\code{whirl_r_session$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="processx" data-topic="process" data-id="as_ps_handle"><a href='../../processx/html/process.html#method-process-as_ps_handle'><code>processx::process$as_ps_handle()</code></a></span></li>
<li><span class="pkg-link" data-pkg="processx" data-topic="process" data-id="format"><a href='../../processx/html/process.html#method-process-format'><code>processx::process$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="processx" data-topic="process" data-id="get_cmdline"><a href='../../processx/html/process.html#method-process-get_cmdline'><code>processx::process$get_cmdline()</code></a></span></li>
<li><span class="pkg-link" data-pkg="processx" data-topic="process" data-id="get_cpu_times"><a href='../../processx/html/process.html#method-process-get_cpu_times'><code>processx::process$get_cpu_times()</code></a></span></li>
<li><span class="pkg-link" data-pkg="processx" data-topic="process" data-id="get_error_connection"><a href='../../processx/html/process.html#method-process-get_error_connection'><code>processx::process$get_error_connection()</code></a></span></li>
<li><span class="pkg-link" data-pkg="processx" data-topic="process" data-id="get_error_file"><a href='../../processx/html/process.html#method-process-get_error_file'><code>processx::process$get_error_file()</code></a></span></li>
<li><span class="pkg-link" data-pkg="processx" data-topic="process" data-id="get_exe"><a href='../../processx/html/process.html#method-process-get_exe'><code>processx::process$get_exe()</code></a></span></li>
<li><span class="pkg-link" data-pkg="processx" data-topic="process" data-id="get_exit_status"><a href='../../processx/html/process.html#method-process-get_exit_status'><code>processx::process$get_exit_status()</code></a></span></li>
<li><span class="pkg-link" data-pkg="processx" data-topic="process" data-id="get_input_connection"><a href='../../processx/html/process.html#method-process-get_input_connection'><code>processx::process$get_input_connection()</code></a></span></li>
<li><span class="pkg-link" data-pkg="processx" data-topic="process" data-id="get_input_file"><a href='../../processx/html/process.html#method-process-get_input_file'><code>processx::process$get_input_file()</code></a></span></li>
<li><span class="pkg-link" data-pkg="processx" data-topic="process" data-id="get_memory_info"><a href='../../processx/html/process.html#method-process-get_memory_info'><code>processx::process$get_memory_info()</code></a></span></li>
<li><span class="pkg-link" data-pkg="processx" data-topic="process" data-id="get_name"><a href='../../processx/html/process.html#method-process-get_name'><code>processx::process$get_name()</code></a></span></li>
<li><span class="pkg-link" data-pkg="processx" data-topic="process" data-id="get_output_connection"><a href='../../processx/html/process.html#method-process-get_output_connection'><code>processx::process$get_output_connection()</code></a></span></li>
<li><span class="pkg-link" data-pkg="processx" data-topic="process" data-id="get_output_file"><a href='../../processx/html/process.html#method-process-get_output_file'><code>processx::process$get_output_file()</code></a></span></li>
<li><span class="pkg-link" data-pkg="processx" data-topic="process" data-id="get_pid"><a href='../../processx/html/process.html#method-process-get_pid'><code>processx::process$get_pid()</code></a></span></li>
<li><span class="pkg-link" data-pkg="processx" data-topic="process" data-id="get_poll_connection"><a href='../../processx/html/process.html#method-process-get_poll_connection'><code>processx::process$get_poll_connection()</code></a></span></li>
<li><span class="pkg-link" data-pkg="processx" data-topic="process" data-id="get_result"><a href='../../processx/html/process.html#method-process-get_result'><code>processx::process$get_result()</code></a></span></li>
<li><span class="pkg-link" data-pkg="processx" data-topic="process" data-id="get_start_time"><a href='../../processx/html/process.html#method-process-get_start_time'><code>processx::process$get_start_time()</code></a></span></li>
<li><span class="pkg-link" data-pkg="processx" data-topic="process" data-id="get_status"><a href='../../processx/html/process.html#method-process-get_status'><code>processx::process$get_status()</code></a></span></li>
<li><span class="pkg-link" data-pkg="processx" data-topic="process" data-id="get_username"><a href='../../processx/html/process.html#method-process-get_username'><code>processx::process$get_username()</code></a></span></li>
<li><span class="pkg-link" data-pkg="processx" data-topic="process" data-id="get_wd"><a href='../../processx/html/process.html#method-process-get_wd'><code>processx::process$get_wd()</code></a></span></li>
<li><span class="pkg-link" data-pkg="processx" data-topic="process" data-id="has_error_connection"><a href='../../processx/html/process.html#method-process-has_error_connection'><code>processx::process$has_error_connection()</code></a></span></li>
<li><span class="pkg-link" data-pkg="processx" data-topic="process" data-id="has_input_connection"><a href='../../processx/html/process.html#method-process-has_input_connection'><code>processx::process$has_input_connection()</code></a></span></li>
<li><span class="pkg-link" data-pkg="processx" data-topic="process" data-id="has_output_connection"><a href='../../processx/html/process.html#method-process-has_output_connection'><code>processx::process$has_output_connection()</code></a></span></li>
<li><span class="pkg-link" data-pkg="processx" data-topic="process" data-id="has_poll_connection"><a href='../../processx/html/process.html#method-process-has_poll_connection'><code>processx::process$has_poll_connection()</code></a></span></li>
<li><span class="pkg-link" data-pkg="processx" data-topic="process" data-id="interrupt"><a href='../../processx/html/process.html#method-process-interrupt'><code>processx::process$interrupt()</code></a></span></li>
<li><span class="pkg-link" data-pkg="processx" data-topic="process" data-id="is_alive"><a href='../../processx/html/process.html#method-process-is_alive'><code>processx::process$is_alive()</code></a></span></li>
<li><span class="pkg-link" data-pkg="processx" data-topic="process" data-id="is_incomplete_error"><a href='../../processx/html/process.html#method-process-is_incomplete_error'><code>processx::process$is_incomplete_error()</code></a></span></li>
<li><span class="pkg-link" data-pkg="processx" data-topic="process" data-id="is_incomplete_output"><a href='../../processx/html/process.html#method-process-is_incomplete_output'><code>processx::process$is_incomplete_output()</code></a></span></li>
<li><span class="pkg-link" data-pkg="processx" data-topic="process" data-id="is_supervised"><a href='../../processx/html/process.html#method-process-is_supervised'><code>processx::process$is_supervised()</code></a></span></li>
<li><span class="pkg-link" data-pkg="processx" data-topic="process" data-id="kill"><a href='../../processx/html/process.html#method-process-kill'><code>processx::process$kill()</code></a></span></li>
<li><span class="pkg-link" data-pkg="processx" data-topic="process" data-id="kill_tree"><a href='../../processx/html/process.html#method-process-kill_tree'><code>processx::process$kill_tree()</code></a></span></li>
<li><span class="pkg-link" data-pkg="processx" data-topic="process" data-id="poll_io"><a href='../../processx/html/process.html#method-process-poll_io'><code>processx::process$poll_io()</code></a></span></li>
<li><span class="pkg-link" data-pkg="processx" data-topic="process" data-id="read_all_error"><a href='../../processx/html/process.html#method-process-read_all_error'><code>processx::process$read_all_error()</code></a></span></li>
<li><span class="pkg-link" data-pkg="processx" data-topic="process" data-id="read_all_error_lines"><a href='../../processx/html/process.html#method-process-read_all_error_lines'><code>processx::process$read_all_error_lines()</code></a></span></li>
<li><span class="pkg-link" data-pkg="processx" data-topic="process" data-id="read_all_output"><a href='../../processx/html/process.html#method-process-read_all_output'><code>processx::process$read_all_output()</code></a></span></li>
<li><span class="pkg-link" data-pkg="processx" data-topic="process" data-id="read_all_output_lines"><a href='../../processx/html/process.html#method-process-read_all_output_lines'><code>processx::process$read_all_output_lines()</code></a></span></li>
<li><span class="pkg-link" data-pkg="processx" data-topic="process" data-id="read_error"><a href='../../processx/html/process.html#method-process-read_error'><code>processx::process$read_error()</code></a></span></li>
<li><span class="pkg-link" data-pkg="processx" data-topic="process" data-id="read_error_lines"><a href='../../processx/html/process.html#method-process-read_error_lines'><code>processx::process$read_error_lines()</code></a></span></li>
<li><span class="pkg-link" data-pkg="processx" data-topic="process" data-id="read_output"><a href='../../processx/html/process.html#method-process-read_output'><code>processx::process$read_output()</code></a></span></li>
<li><span class="pkg-link" data-pkg="processx" data-topic="process" data-id="read_output_lines"><a href='../../processx/html/process.html#method-process-read_output_lines'><code>processx::process$read_output_lines()</code></a></span></li>
<li><span class="pkg-link" data-pkg="processx" data-topic="process" data-id="resume"><a href='../../processx/html/process.html#method-process-resume'><code>processx::process$resume()</code></a></span></li>
<li><span class="pkg-link" data-pkg="processx" data-topic="process" data-id="signal"><a href='../../processx/html/process.html#method-process-signal'><code>processx::process$signal()</code></a></span></li>
<li><span class="pkg-link" data-pkg="processx" data-topic="process" data-id="supervise"><a href='../../processx/html/process.html#method-process-supervise'><code>processx::process$supervise()</code></a></span></li>
<li><span class="pkg-link" data-pkg="processx" data-topic="process" data-id="suspend"><a href='../../processx/html/process.html#method-process-suspend'><code>processx::process$suspend()</code></a></span></li>
<li><span class="pkg-link" data-pkg="processx" data-topic="process" data-id="wait"><a href='../../processx/html/process.html#method-process-wait'><code>processx::process$wait()</code></a></span></li>
<li><span class="pkg-link" data-pkg="processx" data-topic="process" data-id="write_input"><a href='../../processx/html/process.html#method-process-write_input'><code>processx::process$write_input()</code></a></span></li>
<li><span class="pkg-link" data-pkg="callr" data-topic="r_session" data-id="attach"><a href='../../callr/html/r_session.html#method-r_session-attach'><code>callr::r_session$attach()</code></a></span></li>
<li><span class="pkg-link" data-pkg="callr" data-topic="r_session" data-id="call"><a href='../../callr/html/r_session.html#method-r_session-call'><code>callr::r_session$call()</code></a></span></li>
<li><span class="pkg-link" data-pkg="callr" data-topic="r_session" data-id="close"><a href='../../callr/html/r_session.html#method-r_session-close'><code>callr::r_session$close()</code></a></span></li>
<li><span class="pkg-link" data-pkg="callr" data-topic="r_session" data-id="debug"><a href='../../callr/html/r_session.html#method-r_session-debug'><code>callr::r_session$debug()</code></a></span></li>
<li><span class="pkg-link" data-pkg="callr" data-topic="r_session" data-id="get_running_time"><a href='../../callr/html/r_session.html#method-r_session-get_running_time'><code>callr::r_session$get_running_time()</code></a></span></li>
<li><span class="pkg-link" data-pkg="callr" data-topic="r_session" data-id="get_state"><a href='../../callr/html/r_session.html#method-r_session-get_state'><code>callr::r_session$get_state()</code></a></span></li>
<li><span class="pkg-link" data-pkg="callr" data-topic="r_session" data-id="poll_process"><a href='../../callr/html/r_session.html#method-r_session-poll_process'><code>callr::r_session$poll_process()</code></a></span></li>
<li><span class="pkg-link" data-pkg="callr" data-topic="r_session" data-id="read"><a href='../../callr/html/r_session.html#method-r_session-read'><code>callr::r_session$read()</code></a></span></li>
<li><span class="pkg-link" data-pkg="callr" data-topic="r_session" data-id="run"><a href='../../callr/html/r_session.html#method-r_session-run'><code>callr::r_session$run()</code></a></span></li>
<li><span class="pkg-link" data-pkg="callr" data-topic="r_session" data-id="run_with_output"><a href='../../callr/html/r_session.html#method-r_session-run_with_output'><code>callr::r_session$run_with_output()</code></a></span></li>
<li><span class="pkg-link" data-pkg="callr" data-topic="r_session" data-id="traceback"><a href='../../callr/html/r_session.html#method-r_session-traceback'><code>callr::r_session$traceback()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-whirl_r_session-new"></a>}}
\if{latex}{\out{\hypertarget{method-whirl_r_session-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new whirl R session
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{whirl_r_session$new(
  lib_paths = .libPaths(),
  verbosity_level = zephyr::get_opt("verbosity_level")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lib_paths}}{character vector of library paths}

\item{\code{verbosity_level}}{character string of verbosity level}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An \code{whirl_r_session} object, that inherits from \code{callr::r_session}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-whirl_r_session-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-whirl_r_session-finalize}{}}}
\subsection{Method \code{finalize()}}{
Finalize the whirl R session
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{whirl_r_session$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-whirl_r_session-print"></a>}}
\if{latex}{\out{\hypertarget{method-whirl_r_session-print}{}}}
\subsection{Method \code{print()}}{
Print the whirl R session object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{whirl_r_session$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\link{invisible} self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-whirl_r_session-get_dir"></a>}}
\if{latex}{\out{\hypertarget{method-whirl_r_session-get_dir}{}}}
\subsection{Method \code{get_dir()}}{
Get the directory of the whirl R session
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{whirl_r_session$get_dir()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-whirl_r_session-start_strace"></a>}}
\if{latex}{\out{\hypertarget{method-whirl_r_session-start_strace}{}}}
\subsection{Method \code{start_strace()}}{
Start strace
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{whirl_r_session$start_strace()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-whirl_r_session-run_something"></a>}}
\if{latex}{\out{\hypertarget{method-whirl_r_session-run_something}{}}}
\subsection{Method \code{run_something()}}{
Run
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{whirl_r_session$run_something()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-whirl_r_session-clone"></a>}}
\if{latex}{\out{\hypertarget{method-whirl_r_session-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{whirl_r_session$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
