% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_msg.R
\name{log_msg}
\alias{log_msg}
\alias{log_read}
\alias{log_write}
\alias{log_custom}
\title{Helper function to log custom messages}
\usage{
log_msg(msg, op = c("custom", "read", "write"))

log_read(msg)

log_write(msg)

log_custom(msg)
}
\arguments{
\item{msg}{\code{\link[=character]{character()}} Message to display in the log}

\item{op}{\code{\link[=character]{character()}} Operation that was performed}
}
\description{
Useful for e.g. read and write operations on databases etc.
that are not automatically captured.
}
