% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_by_config.R
\name{logging_from_yaml}
\alias{logging_from_yaml}
\title{Read and create summary data frame}
\usage{
logging_from_yaml(file, summary_dir, root_dir, cli_level = cli::cli_h2)
}
\arguments{
\item{file}{configuration to define steps, have to be a yaml}

\item{summary_dir}{A character string of file path specifying the directory
where the summary log will be stored.}

\item{root_dir}{the root dir of the yaml file}

\item{cli_level}{Have to be a function from cli}
}
\description{
Read and create summary data frame
}
