% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{options}
\alias{options}
\title{whirl Options}
\description{
Internally used, package-specific options. All options will prioritize R options() values, and fall back to environment variables if undefined. If neither the option nor the environment variable is set, a default value is used.
}
\section{Checking Option Values}{

Option values specific to \code{whirl} can be
accessed by passing the package name to \code{env}.

\if{html}{\out{<div class="sourceCode">}}\preformatted{options::opts(env = "whirl")

options::opt(x, default, env = "whirl")
}\if{html}{\out{</div>}}
}

\section{Options}{

\describe{
\item{track_files}{\describe{
Should files read and written be tracked? Currently only supported on Linux.\item{default: }{\preformatted{FALSE}}
\item{option: }{whirl.track_files}
\item{envvar: }{R_WHIRL_TRACK_FILES (TRUE if one of 'TRUE', '1', FALSE otherwise)}
}}

\item{check_renv}{\describe{
Should the projects renv status be checked?\item{default: }{\preformatted{FALSE}}
\item{option: }{whirl.check_renv}
\item{envvar: }{R_WHIRL_CHECK_RENV (TRUE if one of 'TRUE', '1', FALSE otherwise)}
}}

\item{track_files_discards}{\describe{
List of file naming patterns not be tracked when track_files = TRUE\item{default: }{\preformatted{c("/lib", "/etc", "/lib64", "/usr", "/var", "/opt", "/sys", "/proc", 
    "/tmp", "/.$", "/null", "/urandom", "/.cache", "/renv", .libPaths())}}
\item{option: }{whirl.track_files_discards}
\item{envvar: }{R_WHIRL_TRACK_FILES_DISCARDS (as character vector, split on ';' delimiter)}
}}

}
}

\seealso{
options getOption Sys.setenv Sys.getenv
}
