% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readstrace_info.R
\name{readstrace_info}
\alias{readstrace_info}
\title{Get session info}
\usage{
readstrace_info(path)
}
\arguments{
\item{path}{a character vector with path name}
}
\description{
Retrieve session info and add quarto info if not already there
}
