% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_summary.R
\name{render_summary}
\alias{render_summary}
\title{Render dataframe into a summary.html file}
\usage{
render_summary(input, summary_dir)
}
\arguments{
\item{input}{The input data.frame that should be rendered into a summary.html file}

\item{summary_dir}{The output directory}
}
\description{
Render dataframe into a summary.html file
}
