% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{retrieve_fpath}
\alias{retrieve_fpath}
\title{Retrieve path of file from folder/subfolders}
\usage{
retrieve_fpath(pat)
}
\arguments{
\item{pat}{\itemize{
\item file to search for
}}
}
\value{
A path to a file
}
\description{
Retrieve path of file from folder/subfolders
}
\examples{
retrieve_fpath("prog1.R")
}
