% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{Execute single or multiple R, R Markdown, and Quarto scripts}
\usage{
run(
  input,
  steps = NULL,
  n_workers = options::opt("n_workers", env = "whirl"),
  summary_file = "summary.html"
)
}
\arguments{
\item{input}{A character vector of file path(s) to R, R Markdown, Quarto
scripts, or files in a folder using regular expression, or to to a whirl
config file. The input can also be structured in a list where each element
will be executed sequentially, while scripts within each element can be
executed in parallel.}

\item{steps}{An optional argument that can be used if only certain steps
within a config files (or list) is to be executed. Should be equivalent to
the names of the steps found in the config file. If kept as NULL (default)
then all steps listed in the config file will be executed.}

\item{n_workers}{Number of simultanous workers used in the run function (Defaults to \code{1}, overwritable using option 'whirl.n_workers' or environment variable 'R_WHIRL_N_WORKERS')}

\item{summary_file}{A character string specifying the file path where the
summary log will be stored.}
}
\value{
A tibble containing the execution results for all the scripts.
}
\description{
Executes and logs the execution of the scripts.
Logs for each script are stored in the same folder as the script.

The way the execution is logged is configurable through several options for
e.g. the verbosity of the logs. See \link{options} on how to configure these.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

# Run a single script
script <- system.file("examples/simple/success.R", package = "whirl")
run(script)

# Run several scripts in parallel on up to 2 workers
scripts <- system.file("examples/simple", c("success.R", "warning.R", "error.R"), package = "whirl")
run(scripts, n_workers = 2)

# Run scripts in several steps
step_1 <- system.file("examples/simple", c("success.R", "warning.R"), package = "whirl")
step_2 <- system.file("examples/simple", c("error.R"), package = "whirl")
run(list(step_1, step_2), n_workers = 2)
\dontshow{\}) # examplesIf}
}
