% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{Execute single or multiple R, R Markdown, and Quarto scripts}
\usage{
run(
  input,
  steps = NULL,
  summary_file = "summary.html",
  n_workers = options::opt("n_workers", env = "whirl"),
  check_renv = options::opt("check_renv", env = "whirl"),
  verbosity_level = options::opt("verbosity_level", env = "whirl"),
  track_files = options::opt("track_files", env = "whirl"),
  out_formats = options::opt("out_formats", env = "whirl")
)
}
\arguments{
\item{input}{A character vector of file path(s) to R, R Markdown, Quarto
scripts, or files in a folder using regular expression, or to to a whirl
config file. The input can also be structured in a list where each element
will be executed sequentially, while scripts within each element can be
executed in parallel.}

\item{steps}{An optional argument that can be used if only certain steps
within a config files (or list) is to be executed. Should be equivalent to
the names of the steps found in the config file. If kept as NULL (default)
then all steps listed in the config file will be executed.}

\item{summary_file}{A character string specifying the file path where the
summary log will be stored.}

\item{n_workers}{Number of simultanous workers used in the run function. A maximum of 128 workers is allowed. (Defaults to \code{1}, overwritable using option 'whirl.n_workers' or environment variable 'R_WHIRL_N_WORKERS')}

\item{check_renv}{Should the projects renv status be checked? (Defaults to \code{FALSE}, overwritable using option 'whirl.check_renv' or environment variable 'R_WHIRL_CHECK_RENV')}

\item{verbosity_level}{How chatty should the log be? Possibilities are \code{quiet}, \code{minimal} and \code{verbose}. (Defaults to \code{"verbose"}, overwritable using option 'whirl.verbosity_level' or environment variable 'R_WHIRL_VERBOSITY_LEVEL')}

\item{track_files}{Should files read and written be tracked? Currently only supported on Linux. (Defaults to \code{FALSE}, overwritable using option 'whirl.track_files' or environment variable 'R_WHIRL_TRACK_FILES')}

\item{out_formats}{Which log format(s) to produce. Possiblities are \code{html}, \code{json}, and markdown formats:\code{gfm}, \code{commonmark}, and \code{markua}. (Defaults to \code{"html"}, overwritable using option 'whirl.out_formats' or environment variable 'R_WHIRL_OUT_FORMATS')}
}
\value{
A tibble containing the execution results for all the scripts.
}
\description{
Executes and logs the execution of the scripts.
Logs for each script are stored in the same folder as the script.

The way the execution is logged is configurable through several options for
e.g. the verbosity of the logs. See \link{options} on how to configure these.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

# Run a single script
script <- system.file("examples/success.R", package = "whirl")
run(script)

# Run several scripts in parallel on up to 2 workers
scripts <- system.file("examples", c("success.R", "warning.R", "error.R"), package = "whirl")
run(scripts, n_workers = 2)

# Run scripts in several steps
step_1 <- system.file("examples", c("success.R", "warning.R"), package = "whirl")
step_2 <- system.file("examples", c("error.R"), package = "whirl")
run(list(step_1, step_2), n_workers = 2)
\dontshow{\}) # examplesIf}
}
