% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{run_script}
\alias{run_script}
\title{Run script
test with a single script}
\usage{
run_script(
  script,
  track_files = FALSE,
  strace_discards = NULL,
  renv = FALSE,
  output_format = "gfm",
  session_info_json = FALSE,
  out_dir = dirname(script)
)
}
\arguments{
\item{script}{path}

\item{track_files}{logical}

\item{strace_discards}{keywords to use to discard not required lines}

\item{renv}{logical}

\item{output_format}{format of the log output in addition to html}

\item{session_info_json}{session_info in json format}

\item{out_dir}{description}
}
\description{
Run script
test with a single script
}
