% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{run_script}
\alias{run_script}
\title{Run script
test with a single script}
\usage{
run_script(
  script,
  track_files = options::opt("track_files"),
  check_renv = options::opt("check_renv"),
  out_dir = dirname(script)
)
}
\arguments{
\item{script}{path}

\item{track_files}{Should files read and written be tracked? Currently only supported on Linux. (Defaults to \code{FALSE}, overwritable using option 'whirl.track_files' or environment variable 'R_WHIRL_TRACK_FILES')}

\item{check_renv}{Should the projects renv status be checked? (Defaults to \code{FALSE}, overwritable using option 'whirl.check_renv' or environment variable 'R_WHIRL_CHECK_RENV')}

\item{out_dir}{description}
}
\description{
Run script
test with a single script
}
