% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{run_script}
\alias{run_script}
\title{Run script
test with a single script}
\usage{
run_script(
  script,
  track_files = FALSE,
  strace_discards = NULL,
  renv = FALSE,
  out_format = "html",
  out_dir = dirname(script)
)
}
\arguments{
\item{script}{path The .R/qmd/Rmd files to be executed}

\item{track_files}{logical}

\item{strace_discards}{keywords to use to discard not required lines}

\item{renv}{logical}

\item{out_format}{format of the log output either html/md}

\item{out_dir}{description}
}
\description{
Run script
test with a single script
}
