% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{run_script}
\alias{run_script}
\title{Run script
test with a single script}
\usage{
run_script(
  script,
  strace = FALSE,
  renv = TRUE,
  cleanup = TRUE,
  output_dir = NULL
)
}
\arguments{
\item{script}{path}

\item{strace}{logical}

\item{renv}{logical}

\item{cleanup}{logical}
}
\description{
Run script
test with a single script
}
