% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{run_script}
\alias{run_script}
\title{Run script
test with a single script}
\usage{
run_script(script, track_files = FALSE, renv = TRUE, out_dir = dirname(script))
}
\arguments{
\item{script}{path}

\item{renv}{logical}

\item{out_dir}{description}

\item{strace}{logical}
}
\description{
Run script
test with a single script
}
