% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{whirl-options}
\alias{whirl-options}
\title{Whirl Package Options}
\description{
Internally used, package-specific options for whirl. All options will prioritize
R options() values, and fall back to environment variables if undefined. If
neither the option nor the environment variable is set, a default value is used.
}
\section{Checking Option Values}{

Option values specific to \code{whirl} can be accessed by passing the package name to \code{env}.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{zephyr::opts_pkg(env = "whirl")
zephyr::opt_pkg(x, default, env = "whirl")
}\if{html}{\out{</div>}}
}

\section{Options}{

\describe{
\item{\code{approved_pkgs_folder}}{
Approved folder library packages
\itemize{
\item Default: NULL
\item Option: \code{whirl.approved_pkgs_folder}
\item Environment variable: \code{R_APPROVED_PKGS_FOLDER} (as character vector, split on ';' delimiter)
}
}
\item{\code{approved_pkgs_url}}{
Approved URL library packages
\itemize{
\item Default: NULL
\item Option: \code{whirl.approved_pkgs_url}
\item Environment variable: \code{R_APPROVED_PKGS_URL} (as character vector, split on ';' delimiter)
}
}
\item{\code{check_renv}}{
Should the projects renv status be checked?
\itemize{
\item Default: \preformatted{
FALSE
}
\item Option: \code{whirl.check_renv}
\item Environment variable: \code{R_CHECK_RENV} (TRUE if one of 'TRUE', '1', FALSE otherwise)
}
}
\item{\code{n_workers}}{
Number of simultanous workers used in the run function. A maximum of 128
workers is allowed.
\itemize{
\item Default: \preformatted{
1
}
\item Option: \code{whirl.n_workers}
\item Environment variable: \code{R_N_WORKERS} (evaluated if possible, raw string otherwise)
}
}
\item{\code{out_formats}}{
Which log format(s) to produce. Possiblities are \code{html}, \code{json}, and markdown
formats:\code{gfm}, \code{commonmark}, and \code{markua}.
\itemize{
\item Default: \preformatted{
"html"
}
\item Option: \code{whirl.out_formats}
\item Environment variable: \code{R_OUT_FORMATS} (evaluated if possible, raw string otherwise)
}
}
\item{\code{track_files}}{
Should files read and written be tracked? Currently only supported on Linux.
\itemize{
\item Default: \preformatted{
FALSE
}
\item Option: \code{whirl.track_files}
\item Environment variable: \code{R_TRACK_FILES} (TRUE if one of 'TRUE', '1', FALSE otherwise)
}
}
\item{\code{track_files_discards}}{
List of file naming patterns not be tracked when track_files = TRUE
\itemize{
\item Default: \preformatted{
c("^/lib", "^/etc", "^/lib64", "^/usr", "^/var", "^/opt", "^/sys",
  "^/proc", "^/tmp", "^/null", "^/urandom", "^/.cache",
  "/scer/homedirs/lvgk/R/x86_64-pc-linux-gnu-library/4.4.1",
  "/opt/R/4.4.1/lib/R/site-library", "/opt/R/4.4.1/lib/R/library")
}
\item Option: \code{whirl.track_files_discards}
\item Environment variable: \code{R_TRACK_FILES_DISCARDS} (evaluated if possible, raw string otherwise)
}
}
\item{\code{track_files_keep}}{
List of file naming patterns alway to be tracked when track_files = TRUE
\itemize{
\item Default: \preformatted{
"^/scer/homedirs/lvgk/whirl"
}
\item Option: \code{whirl.track_files_keep}
\item Environment variable: \code{R_TRACK_FILES_KEEP} (evaluated if possible, raw string otherwise)
}
}
\item{\code{verbosity_level}}{
How chatty should the log be? Possibilities are \code{quiet}, \code{minimal} and
\code{verbose}.
\itemize{
\item Default: \preformatted{
"verbose"
}
\item Option: \code{whirl.verbosity_level}
\item Environment variable: \code{R_VERBOSITY_LEVEL} (evaluated if possible, raw string otherwise)
}
}
}
}

\seealso{
\code{\link[=options]{options()}}, \code{\link[=getOption]{getOption()}}, \code{\link[=Sys.setenv]{Sys.setenv()}}, \code{\link[=Sys.getenv]{Sys.getenv()}}
}
