% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whirl-package.R
\docType{package}
\name{whirl-package}
\alias{whirl}
\alias{whirl-package}
\title{whirl: Loggin package}
\description{
Defines common standards, workflows, and actions for new R packages.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

Other contributors:
\itemize{
  \item Novo Nordisk A/S [copyright holder]
}

}
\keyword{internal}
