% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whirl-package.R
\docType{package}
\name{whirl-package}
\alias{whirl}
\alias{whirl-package}
\title{whirl: Logging package}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides functionalities for running R scripts in batch, while simultanously creating logs for each script execution.
}
\author{
\strong{Maintainer}: Aksel Thomsen \email{oath@novonordisk.com}

Authors:
\itemize{
  \item Lovemore Gakava \email{lvgk@novonordisk.com}
  \item Cervan Girard \email{cgid@novonordisk.com}
  \item Kristian Troejelsgaard \email{ktqn@novonordisk.com}
}

Other contributors:
\itemize{
  \item Novo Nordisk A/S [copyright holder]
}

}
\keyword{internal}
