% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whirl-package.R
\docType{package}
\name{whirl-package}
\alias{whirl}
\alias{whirl-package}
\title{whirl: Logging package}
\description{
Defines common standards, workflows, and actions for new R packages.
}
\author{
\strong{Maintainer}: Aksel Thomsen \email{oath@novonordisk.com}

Authors:
\itemize{
  \item Lovemore Gakava \email{lvgk@novonordisk.com}
  \item Cervan Girard \email{cgid@novonordisk.com}
}

Other contributors:
\itemize{
  \item Novo Nordisk A/S [copyright holder]
}

}
\keyword{internal}
