% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whirl-package.R
\docType{package}
\name{whirl-package}
\alias{whirl}
\alias{whirl-package}
\title{whirl: Execute and log execution of scripts}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Execute scripts in batch, while simultaneously creating logs for each script execution.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://novonordisk-opensource.github.io/whirl/}
  \item \url{https://github.com/novonordisk-opensource/whirl}
  \item Report bugs at \url{https://github.com/NovoNordisk-OpenSource/whirl/issues}
}

}
\author{
\strong{Maintainer}: Aksel Thomsen \email{oath@novonordisk.com}

Authors:
\itemize{
  \item Lovemore Gakava \email{lvgk@novonordisk.com}
  \item Cervan Girard \email{cgid@novonordisk.com}
  \item Kristian Troejelsgaard \email{ktqn@novonordisk.com}
  \item Steffen Falgreen Larsen \email{sffl@novonordisk.com}
  \item Vladimir Obucina \email{vlob@novonordisk.com}
}

Other contributors:
\itemize{
  \item Novo Nordisk A/S [copyright holder]
}

}
\keyword{internal}
