% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocompute.R
\name{write_biocompute}
\alias{write_biocompute}
\title{Create biocompute logs}
\usage{
write_biocompute(queue = run("_whirl.yml"), path = "bco.json", ...)
}
\arguments{
\item{queue}{Result from \code{run()}}

\item{path}{description}

\item{...}{Additional arguments parsed to \code{jsonlite::write_json()}. Note always uses \code{auto_unbox = TRUE}.}
}
\value{
(\code{ìnvisible}) \code{list} of the biocompute domains and their content.`
}
\description{
biocompute stuff
}
